+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 - Releases Firestore Emulator v1.12.0: supports clearing data partially.
 - Fixes manually setting download tokens in Storage Emulator. (#3396)
 - Fixes deleting custom metadata in Storage emulator. (#3385)
+- Fixes errors when calling makePublic() with Storage Emulator(#3394)